Separated namespaces, using filegenerators, using outputformats

+++ src/DocsGenerator/Program.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.IO;
-using System.Reflection;
+using System.Runtime.InteropServices;
 using McMaster.Extensions.CommandLineUtils;
-using DocsGenerator.Markdown;
-using System.Linq;
-using DocsGenerator.Utils;
-using DocsGenerator.Extensions;
+using DocsGenerator.Tools;
 
 namespace DocsGenerator
 {
@@ -13,6 +9,13 @@ namespace DocsGenerator
     {
         public static int Main(string[] args)
         {
+            CommandOption debug = null;
+
+            // Console.WriteLine($"Runtime FrameworkDescription: {RuntimeInformation.FrameworkDescription}");
+            // Console.WriteLine($"Runtime OSArchitecture: {RuntimeInformation.OSArchitecture}");
+            // Console.WriteLine($"Runtime OSDescription: {RuntimeInformation.OSDescription}");
+            // Console.WriteLine($"Runtime ProcessArchitecture: {RuntimeInformation.ProcessArchitecture}");
+
             try
             {
                 var app = new CommandLineApplication();
@@ -20,21 +23,28 @@ namespace DocsGenerator
                 app.HelpOption(inherited: true);
                 var sourcePath = app.Argument("[source]", "The path of the source folder containing dlls to generate documentation for");
                 var outputPath = app.Argument("[output]", "The path of the output folder where files will be generated.");
-                var verbose = app.Option("-v|--verbose", "Flag to enable verbose logging.", CommandOptionType.NoValue);
+                
+                debug = app.Option("-d|--debug", "Flag to enable debug session, which does not delete temporary files.", CommandOptionType.NoValue);
+                var outputFormatOptions = app.Option("-o|--output-format <FORMAT>", "Adds an output format for generating output.", CommandOptionType.MultipleValue);
                 var projectName = app.Option<string>("-n|--name <NAME>", "The name of the project.", CommandOptionType.SingleValue);
+                var verbose = app.Option("-v|--verbose", "Flag to enable verbose logging.", CommandOptionType.NoValue);
             
                 app.OnExecute(() =>
                 {
-                    var settings = new DocsProjectSettings()
+                    var formats = outputFormatOptions.Values;
+
+                    var settings = new ProjectSettings()
                         .WithProjectName(projectName.Value())
                         .HasSource(sourcePath.Value)
                         .HasOutputFolder(outputPath.Value)
-                        .UseMarkdown()
                         .UseVerboseLogging(verbose.HasValue());       
 
-                    Console.WriteLine("Verbose Logging is {0}", verbose.HasValue());                 
+                    if (formats.Contains("markdown")) settings.UseMarkdown();
+                    if (formats.Contains("html")) settings.UseHtml();
 
-                    var builder = new DocsProjectBuilder(settings)
+                    Console.WriteLine("Verbose Logging is {0}", verbose.HasValue());
+
+                    var builder = new ProjectBuilder(settings)
                         .GenerateProject()
                         .Compile();
 
@@ -50,7 +60,7 @@ namespace DocsGenerator
             }
             finally
             {   
-                Console.WriteLine("Directory.Delete(settings.TempFolder, true);");
+                if (!debug.HasValue())Console.WriteLine("Directory.Delete(settings.TempFolder, true);");
                 //Directory.Delete(settings.TempFolder, true);
             }
         }
